Remove timer from syncLoop -- it's now always one-off
We'll be listening to websocket for updates in an upcoming commit, so there is no need for a timer anymore.

syncLoop will now just be a single-shot function that encapsulates the wallet-unlocking (doGetSyncDesktop) and doGetSync, along with updating preferences.

Could be renamed `doFullSync` or something, but retaining the name for now.