Remove "sync on tab-focus"
Probably won't be needing this after listening to websocket, as the time window for tab to be very stale should be minimized. This should also fix the problem of settings being reverted when changed during the tab-focus sync.

Let's see how it goes. Technically it could still fail if the websocket got disconnected in a tab or something.